+++ projects/igniteui-angular-i18n/src/i18n/DE/combo-resources.ts
@@ -2,7 +2,9 @@ import { IComboResourceStrings } from 'igniteui-angular';
 
 // exported below as re-cast to create declaration type with expanded properties
 const ComboResourceStringsDE_: ExpandRequire<IComboResourceStrings> = {
-    igx_combo_empty_message: 'Die Liste ist leer'
+    igx_combo_empty_message: 'Die Liste ist leer',
+    igx_combo_enabledFilter_placeholder: 'Suchbegriff eingeben',
+    igx_combo_addCustomValues_placeholder: 'Element hinzufügen'
 };
 
 /**