+++ projects/igniteui-angular-i18n/src/i18n/ES/combo-resources.ts
@@ -2,7 +2,9 @@ import { IComboResourceStrings } from 'igniteui-angular';
 
 // exported below as re-cast to create declaration type with expanded properties
 const ComboResourceStringsES_: ExpandRequire<IComboResourceStrings> = {
-    igx_combo_empty_message: 'La lista está vacía'
+    igx_combo_empty_message: 'La lista está vacía',
+    igx_combo_enabledFilter_placeholder: 'Escriba un término de búsqueda',
+    igx_combo_addCustomValues_placeholder: 'Agregar elemento'
 };
 
 /**