+++ projects/igniteui-angular-i18n/src/i18n/FR/combo-resources.ts
@@ -2,7 +2,9 @@ import { IComboResourceStrings } from 'igniteui-angular';
 
 // exported below as re-cast to create declaration type with expanded properties
 const ComboResourceStringsFR_: ExpandRequire<IComboResourceStrings> = {
-    igx_combo_empty_message: 'La liste est vide'
+    igx_combo_empty_message: 'La liste est vide',
+    igx_combo_enabledFilter_placeholder: 'Entrez un terme de recherche',
+    igx_combo_addCustomValues_placeholder: 'Ajouter un élément'
 };
 
 /**