+++ projects/igniteui-angular-i18n/src/i18n/HU/combo-resources.ts
@@ -2,7 +2,9 @@ import { IComboResourceStrings } from 'igniteui-angular';
 
 // exported below as re-cast to create declaration type with expanded properties
 const ComboResourceStringsHU_: ExpandRequire<IComboResourceStrings> = {
-    igx_combo_empty_message: 'Üres a lista'
+    igx_combo_empty_message: 'Üres a lista',
+    igx_combo_enabledFilter_placeholder: 'Írjon be egy keresési kifejezést',
+    igx_combo_addCustomValues_placeholder: 'Elem hozzáadása'
 };
 
 /**