+++ projects/igniteui-angular-i18n/src/i18n/IT/combo-resources.ts
@@ -2,7 +2,9 @@ import { IComboResourceStrings } from 'igniteui-angular';
 
 // exported below as re-cast to create declaration type with expanded properties
 const ComboResourceStringsIT_: ExpandRequire<IComboResourceStrings> = {
-    igx_combo_empty_message: 'L\'elenco è vuoto'
+    igx_combo_empty_message: 'L\'elenco è vuoto',
+    igx_combo_enabledFilter_placeholder: 'Immettere il testo di ricerca',
+    igx_combo_addCustomValues_placeholder: 'Aggiungi elemento'
 };
 
 /**