+++ projects/igniteui-angular-i18n/src/i18n/JA/combo-resources.ts
@@ -2,7 +2,9 @@ import { IComboResourceStrings } from 'igniteui-angular';
 
 // exported below as re-cast to create declaration type with expanded properties
 const ComboResourceStringsJA_: ExpandRequire<IComboResourceStrings> = {
-    igx_combo_empty_message: 'リストが空です'
+    igx_combo_empty_message: 'リストが空です',
+    igx_combo_enabledFilter_placeholder: '検索条件を入力',
+    igx_combo_addCustomValues_placeholder: 'アイテムを追加'
 };
 
 /**