+++ projects/igniteui-angular-i18n/src/i18n/KO/combo-resources.ts
@@ -2,7 +2,9 @@ import { IComboResourceStrings } from 'igniteui-angular';
 
 // exported below as re-cast to create declaration type with expanded properties
 const ComboResourceStringsKO_: ExpandRequire<IComboResourceStrings> = {
-    igx_combo_empty_message: '목록이 비어 있음'
+    igx_combo_empty_message: '목록이 비어 있음',
+    igx_combo_enabledFilter_placeholder: '검색어 입력',
+    igx_combo_addCustomValues_placeholder: '항목 추가'
 };
 
 /**