+++ projects/igniteui-angular-i18n/src/i18n/NB/combo-resources.ts
@@ -2,7 +2,9 @@ import { IComboResourceStrings } from 'igniteui-angular';
 
 // exported below as re-cast to create declaration type with expanded properties
 const ComboResourceStringsNB_: ExpandRequire<IComboResourceStrings> = {
-    igx_combo_empty_message: 'Listen er tom'
+    igx_combo_empty_message: 'Listen er tom',
+    igx_combo_enabledFilter_placeholder: 'Skriv inn søkeord',
+    igx_combo_addCustomValues_placeholder: 'Legg til element'
 };
 
 /**