+++ projects/igniteui-angular-i18n/src/i18n/NL/combo-resources.ts
@@ -2,7 +2,9 @@ import { IComboResourceStrings } from 'igniteui-angular';
 
 // exported below as re-cast to create declaration type with expanded properties
 const ComboResourceStringsNL_: ExpandRequire<IComboResourceStrings> = {
-    igx_combo_empty_message: 'De lijst is leeg'
+    igx_combo_empty_message: 'De lijst is leeg',
+    igx_combo_enabledFilter_placeholder: 'Typ een zoekterm',
+    igx_combo_addCustomValues_placeholder: 'Item toevoegen'
 };
 
 /**