+++ projects/igniteui-angular-i18n/src/i18n/PL/combo-resources.ts
@@ -2,7 +2,9 @@ import { IComboResourceStrings } from 'igniteui-angular';
 
 // exported below as re-cast to create declaration type with expanded properties
 const ComboResourceStringsPL_: ExpandRequire<IComboResourceStrings> = {
-    igx_combo_empty_message: 'Lista jest pusta'
+    igx_combo_empty_message: 'Lista jest pusta',
+    igx_combo_enabledFilter_placeholder: 'Wprowadź tekst wyszukiwania',
+    igx_combo_addCustomValues_placeholder: 'Dodaj element'
 };
 
 /**