+++ projects/igniteui-angular-i18n/src/i18n/SV/combo-resources.ts
@@ -2,7 +2,9 @@ import { IComboResourceStrings } from 'igniteui-angular';
 
 // exported below as re-cast to create declaration type with expanded properties
 const ComboResourceStringsSV_: ExpandRequire<IComboResourceStrings> = {
-    igx_combo_empty_message: 'Listan är tom'
+    igx_combo_empty_message: 'Listan är tom',
+    igx_combo_enabledFilter_placeholder: 'Ange sökterm',
+    igx_combo_addCustomValues_placeholder: 'Lägg till objekt'
 };
 
 /**