+++ projects/igniteui-angular-i18n/src/i18n/TR/combo-resources.ts
@@ -2,7 +2,9 @@ import { IComboResourceStrings } from 'igniteui-angular';
 
 // exported below as re-cast to create declaration type with expanded properties
 const ComboResourceStringsTR_: ExpandRequire<IComboResourceStrings> = {
-    igx_combo_empty_message: 'Liste boş'
+    igx_combo_empty_message: 'Liste boş',
+    igx_combo_enabledFilter_placeholder: 'Arama terimi girin',
+    igx_combo_addCustomValues_placeholder: 'Öğe ekle'
 };
 
 /**