+++ projects/igniteui-angular-i18n/src/i18n/ZH-HANT/combo-resources.ts
@@ -2,7 +2,9 @@ import { IComboResourceStrings } from 'igniteui-angular';
 
 // exported below as re-cast to create declaration type with expanded properties
 const ComboResourceStringsZHHANT_: ExpandRequire<IComboResourceStrings> = {
-    igx_combo_empty_message: '清單是空的'
+    igx_combo_empty_message: '清單是空的',
+    igx_combo_enabledFilter_placeholder: '輸入搜尋字串',
+    igx_combo_addCustomValues_placeholder: '新增項目'
 };
 
 /**