+++ projects/igniteui-angular/src/lib/combo/combo.common.ts
@@ -108,8 +108,6 @@ export const enum DataTypes {
 export interface IComboFilteringOptions {
     /** Defines filtering case-sensitivity */
     caseSensitive?: boolean;
-    /** Defines whether filtering is allowed */
-    filterable: boolean;
     /** Defines optional key to filter against complex list items. Default to displayKey if provided.*/
     filteringKey?: string;
 }
@@ -971,7 +969,7 @@ export abstract class IgxComboBaseDirective implements IgxComboBase, AfterViewCh
     private _overlaySettings: OverlaySettings;
     private _groupSortingDirection: SortingDirection = SortingDirection.Asc;
     private _filteringOptions: IComboFilteringOptions;
-    private _defaultFilteringOptions: IComboFilteringOptions = { caseSensitive: false, filterable: true };
+    private _defaultFilteringOptions: IComboFilteringOptions = { caseSensitive: false };
     private _icons = [
         {
             name: 'input_expand',