+++ projects/igniteui-angular/src/lib/core/i18n/combo-resources.ts
@@ -1,7 +1,11 @@
 export interface IComboResourceStrings {
     igx_combo_empty_message?: string;
+    igx_combo_enabledFilter_placeholder?: string;
+    igx_combo_addCustomValues_placeholder?: string;
 }
 
 export const ComboResourceStringsEN: IComboResourceStrings = {
-    igx_combo_empty_message: 'The list is empty'
+    igx_combo_empty_message: 'The list is empty',
+    igx_combo_enabledFilter_placeholder: 'Enter a Search Term',
+    igx_combo_addCustomValues_placeholder: 'Add Item',
 };