+++ CHANGELOG.md
@@ -28,6 +28,9 @@ All notable changes for each version of this project will be documented in this 
     - The `isFirstPageDisabled` and `isLastPageDisabled` have been deprecated in favor of the identical `isFirstPage` and `isLastPage` getter.
 - `IgxOverlayService`
     - The `attach` method overload accepting `Type` and `OverlaySettings` now accepts `OverlayCreateSettings` as second parameter. This interface extends `OverlaySettings` with an additional `injector` property used as `ElementInjector` when creating the dynamic component.
+- `IgxCombo`:
+    - **Breaking Change** The `filterable` property is renamed to `disableFiltering`.
+    - The placeholder of the searchInput is changed depending on the `disableFiltering` option - 'Enter a Search Term' when `disableFiltering` is false and 'Add Item' when `disableFiltering` is true.
 
 ## 18.0.0
 ### New Features